# internal/api/openapi/swagger.yaml
basePath: /api
definitions:
  api.APIError:
    properties:
      code:
        type: integer
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  api.AssignProgramRequest:
    properties:
      goal:
        type: string
      notes:
        type: string
      role:
        example: primary
        type: string
      schedule:
        type: string
      start_date:
        example: "2026-05-12"
        type: string
      template_id:
        example: 7
        type: integer
    type: object
...
  api.AthleteProgram:
    properties:
      active:
        type: boolean
      athlete_id:
        type: integer
      created_at:
        type: string
      deactivated_at:
        type: string
      goal:
        type: string
      id:
        type: integer
      is_loop:
        type: boolean
      notes:
        type: string
      num_days:
        type: integer
      num_weeks:
        type: integer
      role:
        type: string
      schedule:
        type: string
      start_date:
        type: string
      template_id:
        type: integer
      template_name:
        type: string
      updated_at:
        type: string
    type: object
...
  api.BodyWeight:
    properties:
      athlete_id:
        type: integer
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      notes:
        type: string
      weight:
        type: number
    type: object
  api.BodyWeightPage:
    properties:
      entries:
        items:
          $ref: '#/definitions/api.BodyWeight'
        type: array
      has_more:
        type: boolean
    type: object
  api.BodyWeightRequest:
    properties:
      date:
        example: "2026-05-12"
        type: string
      notes:
        type: string
      weight:
        example: 185.5
        type: number
    type: object
...
  api.JournalEntry:
    properties:
      author:
        type: string
      author_id:
        type: integer
      date:
        type: string
      detail:
        type: string
      id:
        type: integer
      is_private:
        type: boolean
      pinned:
        type: boolean
      second_id:
        type: integer
      summary:
        type: string
      type:
        type: string
    type: object
...
  api.ProgramTemplate:
    properties:
      athlete_count:
        type: integer
      athlete_id:
        type: integer
      athlete_name:
        type: string
      audience:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_loop:
        type: boolean
      name:
        type: string
      num_days:
        type: integer
      num_weeks:
        type: integer
      updated_at:
        type: string
    type: object
  api.ProgramTemplateRequest:
    properties:
      athlete_id:
        type: integer
      audience:
        example: adult
        type: string
      description:
        type: string
      is_loop:
        type: boolean
      name:
        example: 5/3/1 BBB
        type: string
      num_days:
        example: 4
        type: integer
      num_weeks:
        example: 3
        type: integer
    type: object
  api.ProgramTemplateUpdateRequest:
    properties:
      description:
        type: string
      is_loop:
        type: boolean
      name:
        example: 5/3/1 BBB
        type: string
      num_days:
        example: 4
        type: integer
      num_weeks:
        example: 3
        type: integer
    type: object
...
  api.StatusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  api.TrainingMax:
    properties:
      athlete_id:
        type: integer
      created_at:
        type: string
      effective_date:
        type: string
      exercise_id:
        type: integer
      exercise_name:
        type: string
      id:
        type: integer
      notes:
        type: string
      weight:
        type: number
    type: object
  api.TrainingMaxRequest:
    properties:
      effective_date:
        example: "2026-05-12"
        type: string
      exercise_id:
        example: 42
        type: integer
      notes:
        type: string
      weight:
        example: 315
        type: number
    type: object
...
info:
  contact: {}
  description: |-
    REST API for the RepLog workout tracking application.

    Authentication uses session cookies issued by `POST /api/login`
    (or the WebAuthn / token-login endpoints). Cookies are
    `HttpOnly`, `SameSite=Lax`, and (in production) `Secure`.
    The browser sends them automatically on subsequent requests.
  title: RepLog API
  version: 1.0.0
paths:
...
  /athletes/{id}/body-weights:
    get:
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BodyWeightPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: List body weights
      tags:
      - Athletes
    post:
      consumes:
      - application/json
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Entry
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.BodyWeightRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.BodyWeight'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Log body weight
      tags:
      - Athletes
  /athletes/{id}/body-weights/{bwID}:
    delete:
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Body weight entry ID
        in: path
        name: bwID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Delete body weight
      tags:
      - Athletes
...
  /athletes/{id}/exercises/{exerciseID}/training-maxes:
    get:
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Exercise ID
        in: path
        name: exerciseID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.TrainingMax'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Get training max history
      tags:
      - TrainingMaxes
...
  /athletes/{id}/journal:
    get:
      description: Combined timeline of workouts, notes, body weights, and goals.
        Coaches see private notes; non-coach linked athletes do not.
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page size (default 50)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.JournalEntry'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Athlete journal timeline
      tags:
      - Athletes
...
  /athletes/{id}/programs:
    get:
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.AthleteProgram'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: List athlete's program assignments
      tags:
      - Programs
    post:
      consumes:
      - application/json
      description: Auto-deactivates any existing active program in the same role (per
        ADR 010).
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Assignment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AssignProgramRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.AthleteProgram'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Assign program to athlete
      tags:
      - Programs
...
  /athletes/{id}/training-maxes:
    get:
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.TrainingMax'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: List current training maxes
      tags:
      - TrainingMaxes
    post:
      consumes:
      - application/json
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: TM
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.TrainingMaxRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.TrainingMax'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Set training max
      tags:
      - TrainingMaxes
...
  /notifications/{notificationID}/read:
    post:
      parameters:
      - description: Notification ID
        in: path
        name: notificationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Mark notification read
      tags:
      - Notifications
...
  /programs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.ProgramTemplate'
            type: array
      summary: List program templates
      tags:
      - Programs
    post:
      consumes:
      - application/json
      parameters:
      - description: Template
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ProgramTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.ProgramTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Create program template
      tags:
      - Programs
  /programs/{id}:
    delete:
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Delete program template
      tags:
      - Programs
    get:
      description: Returns the template metadata plus its prescribed sets.
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Get program template
      tags:
      - Programs
    put:
      consumes:
      - application/json
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      - description: Template
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ProgramTemplateUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ProgramTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Update program template
      tags:
      - Programs
...
swagger: "2.0"
...